prov/sockets: Set av_index in conn_map on passive side
The per endpoint connection map stores a list of active connections.
Each connection stores a av_index that corresponds to that
connection's address in the local AV.

The av_index is left unset (FI_ADDR_NOTAVAIL) if we accept a connection
from a peer.  This can result in mis-matching tagged receive messages,
since FI_ADDR_UNSPEC correponds with any source.

Add a check to sock_ep_lookup_conn() to set the av_index if it is
unset.  Sock_ep_lookup_conn() is called when sending data to a
specific peer, or when receiving a connection message from a peer.
The latter is the first message a peer sends in order to publish its
address.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>