man/progress: Close gap in definitions handling target only EP
Manual progress requires that an app poll the CQ or counter
associated with an endpoint. However, if an endpoint will never
generate completions -- for example, it will only be the target
of RMA operations -- then it is not required to be associated
with any CQ or counter.  This leaves apps with no way to drive
manual progress on the endpoint.

There are a couple of ways to close this gap.  But the only
one that will work with existing providers is to require that
such endpoints be associated with a completion object.  This
will allow the app to drive progress over the endpoint.

Because the man pages state that an endpoint must be associated
with CQs for any types of operations that may be initiated
against the endpoint, we expand that definition to include
endpoints needing manual progress.

This issue showed up in GasNET, which allows for manual progress
on an endpoint dedicated as an RMA target.

Note that as an alternative, we can allow associating an
endpoint with a poll set.  This change doesn't preclude that
as an addition in a future version of the library.  However,
that option is currently not supported by the existing
providers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>