prov/psm2: Fix finalization error in auto progress thread
A list of active Tx/Rx contexts is maintained for each domain. This list
is used by the progress function. Previously there were two issues: for
regular endpoints, the Rx context was added to the list when allocated,
but was not removed from the list when freed, which could cause the auto
progress thread to access invalid context during finalization. For shared
Tx context, the context was not added to the list at all. Both issues
have been addressed here.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>