prov/psm2: Fix incorrect tag received from iov send
As part of the recent changes to support remote CQ date over tagged
messages, the iov protocol was slightly modified to use a different
tag bit layout. As the result, the payload packets for iov sends no
longer carries the original tag information. That means the completion
would have incorrect tag field if generated in the old way.

Update the completion generation routine to handle the new iov protocol.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>