prov/psm2: Improve how Tx/Rx context limits are handled
The number of "free contexts" is volatile and can't be depended on. Check
the user request against the "total contexts" instead. Let the PSM2 library
be the last and only place to determine if a free context is available.

Report the total number of contexts in "fi_info->domain_attr->max_ep_tx_ctx"
and "fi_info->domain_attr->max_ep_rx_ctx". This provides a static view of
the system limit that can actually be used by the application. Report the
current number of free contexts in "fi_info->domain_attr->tx_ctx_cnt" and
"fi_info->domain_attr->rx_ctx_cnt" to serve as a recommended value.

Remove the artificial limit PSMX2_MAX_TRX_CTXT that is no longer used.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>