prov/psm,psm2: Shortcut obviously unsuccessful fi_getinfo() calls
Move some quick checks to the very beginning so that queries that are
apparently unsuitable for the provider will immediately return, thus
avoiding any further steps related to provider initialization.

Also fix multiple invocations of fi_tostr() in a single function call
which could lead to undefined behavior because fi_tostr() uses a single
internal buffer to hold the result.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>