prov/psm2: Improve automatic tag layout selection
Use FI_REMOTE_CQ_DATA in the caps field to distinguish the two provider
instances with or without the remote CQ data support. Also check the
presence of this flag in hints so that the correct provider instance
can be retrieved later by simply setting hints->caps to a previously
returned value. This extends backward compatibility to some applications
that didn't work with the previous method of checking "cq_data_size".

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>