fabric: Introduce variable length message support
Applications often need to receive messages, but have no idea
how large the message will be until the sender posts it.  In
practical terms, the result is that many apps need to
implement a rendezvous protocol using some combination of
tagged messages + RMA, tagged messages, or normal messages + RMA.
This duplicates work that most of the OFI providers already do.
The apps must also guess the best way to do this, which may not
be optimal for a given provider.

This defines a new feature that abstracts the behavior that
applications are wanting -- the ability to receive a message
of variable length, with the size of the message determined
by the sender and given to the receiver as part of a 'pre'
completion event.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>