common: Add delay when using OOB socket
Connecting out of band sockets for address exchange ends up
impacting providers that use an internal name service for
address resolution (mostly done in order to support fabtests).
Because the OOB connection is occuring at the start of any
test that uses it, it halts program execution until the OOB
connection completes.  This causes an issue that the provider
name service cannot start and register local addresses until
both the server and client apps are running.  This ends up
creating a race condition where the client may attempt to
contact the name server prior to it running or having any
addresses registered to it.

When OOB sockets are used insert a delay to allow the server
time to setup its name server.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>