prov/psm2: Fix ep_attr->tx_ctx_cnt/rx_ctx_cnt returned by fi_getinfo
When these fields are not explicitly specified in hints (absent or
with value 0), use default value 1 instead of the number of available
contexts to prevent applications from unintentionally asking for too
many contexts when opening scalable endpoints with default settings.

System resource limits are returned via similar fields in "domain_attr"
so no information is lost.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>