core: Add rbtree derivative, rbmap
Derive a new rbtree from the existing code.  The new rbmap will
differ in a couple of ways.  First, it will only store a data value in
the tree.  The data value will be used to look up the key.
When searching or inserting, a key will be passed in.  The key
will be passed into the compare function along with the stored
data value.   This differs from the rbtree, where two keys
are passed into the compare call.

The other change is that the rbmap structure will will be exposed
to the user.  This will allow it to be embeded directly into other
structures, avoiding extra memory allocations.  The rbmap structure
will also be passed into the compare call for use as a context.

The code is updated to match the style used throughout the rest of
the project.

The use case for rbmap is to allow a provider to maintain a
sorted list of addresses.  However, rather than the addresses
themselves being stored in the tree, an fi_addr from an AV
will be stored.  This reduces the necessary memory footprint,
while providing O(log n) address lookup.  The fi_addr can
be used to lookup the actual address.  Thus, each node will
store the fi_addr value, but the address to search for can be
provided as input into compare.  Providing the rbmap into
the compare calls provides the context needed to associated
an fi_addr with a specific AV.

This will provide a mechanism that utility providers can use
to determine if an address has already been inserted into an
AV, and if so, what the fi_addr is.  Currently, the utility
providers perform an O(n) linear search.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>