prov/util: use a separate flag to determine if we need to create av hash table
This patch introduces a new flag OFI_AV_HASH to determine if we need to create
the hash table for AV. This is to allow providers that require the hash table
but not use FI_SOURCE (if the app doesn't require it).

Also add a new ip_av_create function that accepts flags.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>