prov/sockets: Fix use after free error handling triggered ops
As part of updating a counter, an associated deferred work queue
is checked.  For example, sock_cntr_add() calls
sock_cntr_check_trigger_list() after the counter has been updated.

If an application is waiting on the counter to reach a specific
threshold, it can see that the counter has been set, then attempt
to free the counter, endpoint, and other resources.  This leads
to a condition where the checking of the trigger list may be
called after the counter has been freed.

There's not a great fix for this.  To help avoid potential use
after free issues, hold the counter mutex around the call to
check the trigger list.  Then in destroy acquire and release the
mutex.  This ensures that destroy block until counter updates
have fully completed.

This fixes a crash in MPI as a result of sock_pe_report_write_
completion() calling fi_cntr_addr() when handling a triggered
operation.  MPI frees the counter after fi_cntr_addr() has
incremented the counter, but before the sock_cntr_addr()
call can finish sock_cntr_check_triggere_list().

Signed-off-by: Sean Hefty <sean.hefty@intel.com>