prov/psm2: Avoid unnecessary calls to psmx2_am_progress()
The function is used to process request queues associated with large size
RMA and triggered ops. Previously the initialization of the AM mechanism
was used as the condition to call this function. Since the AM mechanism is
now also used for non-RMA functionality, the condition has become overly
broad and could cause the function be called unnecessarily.

Change the condition to be the presence of FI_RMA or FI_TRIGGERED caps in
the associated endpoints.

Cherry-picked from 45c7bb1c0d9c5adb78afd6b651d6d96b40f724c2.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>