prov/psm2: Avoid unnecessary calls to psmx2_am_progress()
The function is used to process request queues associated with large size
RMA and triggered ops. Previously the initialization of the AM mechanism
was used as the condition to call this function. Since the AM mechanism is
now also used for non-RMA functionality, the condition has become overly
broad and could cause the function be called unnecessarily.

Change the condition to be the presence of FI_RMA or FI_TRIGGERED caps in
the associated endpoints.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>