prov/psm2: Avoid adding FI_TRIGGER caps automatically
FI_TRIGGER is a secondary capability that the provider can optionally
set even if it is not asked for by applications. However, FI_TRIGGER
is related to the calling of psmx2_am_progress() and may affect the
performance. Don't set this caps bit unless it's in the hints or the
hints contain empty caps.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>