prov/rxm: Reduce instruction counts when handling CQ entry
rx_buf->conn is always NULL. rxm_ep_repost_buf sets it to NULL
and issues fi_recv(). When the data (type of message is
ofi_ctrl_data) is arriving into this buffer and CQ is completed,
the function rxm_handle_recv_comp is called to process this buffer.
So, there are no needs to check whether the rx_buf->conn == NULL,
we can just call rxm_key2conn to retrieve connection object.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>