prov/psm2: Fix potential duplication of iov send completion
Under rare condition, it was possible that completions were generated
for both the iov message header and the message body. Fix by counting
the header together with the body before generating completion.

Also make sure that completions for iov sends have tag set in a way
consistent with non-iov sends, even though it is not required by the
API.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>