core/perf: Remove names from ofi_perfset
The perfset structure was defined without an actual user.  Now that
an actual user is being introduced, it turns out that it is simpler
to store the names associated with the counters separately.

Remove the names array from the perfset.  When logging, pass the
counter names in as a separate argument.  This essentially removes
the need for the ofi_perfset_data() call and count field.

The caller is responsible for sizing the perfset appropriately and
associating each counter with an optional name.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>