core: Enable hooking provider calls
Once a fabric has been opened, allow the hook provider to
intercept calls to it.  Hooking is enabled via the FI_HOOK
environment variable.

It should be possible to layer multiple hooks, this patch
limits us to a single hook.  The two defined hooks are
NOOP and PERF.  No-op allows us to verify basic hooking
functionality and is for debug purposes (no documented).
The PERF hook allows us to capture performance data.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>