prov/psm2: Fix incorrect completion event for iov send
A pointer saved for later use to generate iov send completion event could
refer to a variable on the stack and thus could lead to data corruption if
accessed across different invocations of the enclosing function.

Fix by saving the content instead if the pointer refers to a stack variable.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>