prov/util: add a provider specific update function as part of cmap_update
ofi_cmap_update is called on a AV insert call. If we find a handle that's
already in the peer list set a av_updated variable. A provider might need to do
some processing on received messages after a new address is inserted into the
AV. E.g. updating address info on messages in the unexpected queue to correctly\
match pending and future directed recvs.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>