fabric: Introduce new /mode bit FI_BUFFERED_RECV
This bit separates out part of the FI_VARIABLE_MSG feature.
As a mode bit, providers that must provide receive side
buffering can report receive completions to applications
by referencing their buffers directly.  This avoids a
data on the receive side.  The rxm and rxd providers both
can use this function with applications that support the
new mode bit.

Most of the description for FI_BUFFERED_RECV is taken
from the FI_VARIABLE_MSG definition.  However, some
modifications are made to that definition based on
mapping to implementation details.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>