prov/rxm: Fix leak of fi_info
fi_eq.3 man page says:
A connection request (FI_CONNREQ) event indicates that a remote endpoint
wishes to establish a new connection to a listening, or passive, endpoint.
The fid is the passive endpoint. Information regarding the requested,
active endpoint’s capabilities and attributes are available from the info
field. The application is responsible for freeing this structure by
calling fi_freeinfo when it is no longer needed.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>