fi_mr_unspec and ge 1.5 API adjustments
We are trying to bring an OFI BTL online in the Open MPI
and encountering turbulence with the memory registration
model (domain attr mr_mode).

At first we had provider specific code to specify the >= 1.5 API
to keep various providers pacified.  Reviewers did not approve
of this provider specific bit encoding.  So we switched to
using FI_MR_UNSPEC while still using the >= 1.5 API.  Well
that caused problems with several providers, including the GNI
one.

Seems that there are several places in the mr_mode checks in
the utilities code that doesn't take into account the possible
use of FI_MR_UNSPEC while also using the >= 1.5 API.

This PR fixes the checks that cause the case of FI_MR_UNSPEC
to fail.

Otherwise, if this isn't right, what should we be setting the domain
attr mr_mode bits to be in our Open MPI OFI BTL such that it can
work with either fn FI_MR_BASIC or FI_MR_SCALABLE type providers?

Signed-off-by: Howard Pritchard <howardp@lanl.gov>