prov/rxm: Don't overwrite failed return value creating CQ
When running the cq_test over rxm;tcp on windows, the
ep_bind call fails (unable to allocate the tcp cq, since
there's no fd on windows).  However, the failure code is
lost, since it's overwritten when closing the cq on the
error path.

Instead, simply close the cq and return the original error
code.  This avoids the crash.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>