prov/gni: suppress hugepage counting for ARM
Turns out for some reason that on Cray ARM
systems, the output of /proc/self/maps is
not indicating that large page allocations are
being backed by file systems in /var/lib/hugepages/global.

This confuses one of the GNI criterion tests.
Disable the large page counting in these tests
for now.

Signed-off-by: Howard Pritchard <howardp@lanl.gov>