ubertest: redo completions
Because of the addition of cq_bind_flags, op_flags, and the FI_COMPLETION msg_flags, the completions need to be structured to work for every possible path. Important cases that had to be considered:
- tx/rx counters are always incremented, regardless of how the CQ was bound or whether the call was an inject call
- SELECTIVE_COMPLETION + no op_flag + send data -> undefined behavior (skip testing)
- when using both counters and cqs, make sure not to count double completions
- SELECTIVE_COMPLETION + FI_COMPLETION op flags + fi_msg with no flags set -> skip testing (perhaps just for now) - difficult for ubertest to test this specific situation

These issues went under the radar for two reasons, lack of testing (to be dealt with in a follow up commit), and the lack of credit overflow checks (too many completions returning, causing credit tracking to get off). This patch includes asserts for credits after completions are read to make sure we never read too many completions.

Signed-off-by: aingerson <alexia.ingerson@intel.com>