prov/util: Avoid possible crash accessing freed lock
See 1fc1befd7231cc9f054f4131188eb50102d7d102.

While an EP is bound to a CQ, the CQ code path can attempt
to progress transfers on the EP.  This will results in trying
to acquire the EP lock.  In the EP cleanup path, we can't
destroy the EP lock until after it has been unassociated with
the CQ.

Inspect all util objects and move lock destruction to the
end of the cleanup routine to avoid real and potential issues.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>