osd/windows: Rework sendmsg/recvmsg calls
Windows does not support socket send or receive calls that
take an iovec on a streaming socket.  The current abstraction
ends up allocating a buffer, copying the iovec in/out of it,
and using the standard send/recv calls.

Optimize this by separating the sendmsg/recvmsg calls based
on the socket type.  For UDP sockets, we can use WSASendMsg/
WSARecvMsg.  For TCP sockets, we iterate over the iovec.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>