prov/tcp: remove connection manager thread
This patch removes the poll_fd_mgr which is managing the
connections. Current scheme uses a thread to wait for connection
events on sock fds that are added to poll_mgr and handle them. Usage
of thread is unnecessary given that OS handles the actual connections
for the sockets interface. We just need a machanism to report the
events to app when it asks for the events.

So it is logical to have the connection management built around the
event queue, without needing a thread.

In the series of follow-up patches I will add that functionality.

Signed-off-by: Venkata Krishna Nimmagadda <venkata.krishna.nimmagadda@intel.com>

Conflicts:
	prov/tcp/src/tcpx_conn_mgr.c