fabric: Introduce new completion semantic
MPI, for some unknown reason, thought that it would be a good
idea to define an asynchronous operation that doesn't complete
until the message has been copied into an application identified
buffer for that message.  MPI calls this operation MPI_SSend.

In order to support this operation, introduce a new completion
type -- FI_MATCH_COMPLETE.  An operation that uses
FI_MATCH_COMPLETE will not complete until the data has been
placed into an application buffer specifically identified
with the operation.

Add a discussion of the different completion semantics to the
fi_cq man page, and move all definitions there.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>