fabric: Extend buffered/variable length messages for tagged
The current definition for buffered and variable length messages
is insufficient for tagged messages.  It basically requires that
the application do the tag matching, which makes the tagged
interface somewhat irrevelant.  Adjust the definition when using
these features with fi_tagged.

The application is responsible for providing the tags, with
the provider handling the matching and providing the buffering.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>