prov/psm2: Remove stale info from address vector when disconnecting
The "epaddr" info stored in address vector becomes invalid when the
corresponding peer is disconnected. Usually it's not an issue because
the entry is not supposed to be used later. However, if an application
uses fi_cq_readfrom to get the source address of incoming messages,
and multiple senders reuse the same HW address (successively, not
concurrently), the receiver may reuse old entries for new peers and
thus attempt to use the stale information instead of setting up new
connections.

Fix by removing stale "epaddr" information from the address vector
when the peer is disconnected.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>