prov/rxm: Improve Rendezvous protocol
RxM on the receiver side registers the MR for the receiver's I/O vector with
the total length of the I/O vector entries when the LMT request is received.
This patch changes behavior to register the MR for the receiver's I/O vector
with the length received from the LMT request (sender's I/O vector total length).

So, this reduces registration time of the MR in the following case (just an example):

Sender side:
fi_send(buf, 1000); -> fi_cq_read();
Receiver side:
fi_recv(buf, 10000000); -> fi_cq_read();

w/ the patch:
RxM will register MR with length = 1000 on the sender and receiver sides.

w/o the patch:
RxM will register MR with length = 1000 on the sender side and 10000000 -
on the receiver side.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>