common: Changes to epoll implementation to fix potential bugs
current implementation of epoll add/delete/mod directly access the fd array even when the poll call is blocked on it. This would create a race condition between prov and OS accessing the fd array. To resolve this, all the requests from user for add/delete/modfiy fds are serialized via a slist and they signal poll with socket pair. The slist is processed by the wait call when it is unblocked. This would mitigate the race condition.

Signed-off-by: Venkata Krishna Nimmagadda <venkata.krishna.nimmagadda@intel.com>