util_buf.c: add support for huge pages on Linux
Use mmap() on Linux to allocate buf pool memory regions in huge pages
when the initial buf pool allocation is greater than or equal to the
system default huge page size. Fallback on ofi_memalign() if determining
the huge page size or the mmap fails for any reason.  Add stubs for
other platforms to add huge page support for those in the future if
needed.

Fixes #4421

Signed-off-by: Robert Wespetal <wesper@amazon.com>