prov/verbs, util/mem_monitor: Fix notofication mechanism for MR cache
The commit fixes notification mechanism provided by Verbs.
The issue was that if user allocates large buffer (ptr:size) and issues
fi_mr_reg() for the segments (ptr1:size1, ptr2:size2, ..., ptrN:sizeN)
of this buffer. And then calls free() for large buffer, the segements won't
be released (The notification will be generated only for the first segment,
if ptr = ptr1).

This commit replaces UT Hash by Red-Black Tree for searching regions for
which the notifier should generate events.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>