prov/cxi: Fix CQ poll when enabling RX context PTE
When an RX context is enabled, a PTE is allocated and enabled. The PTE is
configured with an EQ backing the libfabric receive CQ. This EQ could have
unrelated events on it. Poll the EQ regularly so any unrelated events are
handled correctly while waiting for a state change event.

For NETCASSINI-907.

Signed-off-by: Zach Tiffany <ztiffany@cray.com>