fabric: Add tostr function to base fid
This is needed in order to properly print the new fid_nic
structure, specifically the prov_attr.  But being able to
print internal details about any fid could be a useful
debugging tool, even though the output will be provider
specific.

Expand the base fi_ops with a tostr function and update
the main fi_tostr() function to handle it.

This replaces the FI_TYPE_DEVICE_ATTR tostr()
functionality.  Note that this value has never been
part of a release, so it is safe to remove.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>