prov/rxm: Fix handling of truncated messages in Rendezvous protocol
This patch fixes case when the length of the supplied buffer is smaller than
the length of the message that will be read by fi_readv().

w/a this fix:
Reports -FI_ETRUNC, but doesn't read the data to the supplied buffer
v/ this fix:
Read the data to the supplied buffer and writes -FI_ETRUNC error to
CQ error entry

The problem was reproduced with MPICH errors/pt2pt/truncmsg1 test.

Signed-off-by: Dmitry Gladkov <dmitry.gladkov@intel.com>